sales_coach_v3/digital-sales-coach-voice-first/digital-sales-coach-voice-app/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
gunicorn>=21.0.0
